+++ layla/dbmanager/drivers/mysql.php
@@ -1,28 +1,27 @@
-<?php namespace DBManager\Drivers; 
+<?php namespace Layla\DBManager\Drivers; 
 
 use DBManager;
 use Laravel\Config;
 use Exception;
 use PDO;
 
-class SQLite extends Driver {
+class MySQL extends Driver {
 
 	/**
 	 * Driver specific hidden tables
 	 *
 	 * @var array
 	 **/
-	public static $driver_hidden = array('sqlite_sequence');
+	public static $driver_hidden = array();
 
 	/**
 	 * Return all tables we're allowed to use
 	 *
 	 * @return array
-	 * // pgsql:  $query_string = "SELECT tablename FROM pg_tables WHERE tablename !~ '^pg_+' AND tableowner = '" . $connection['username'] ."'";
 	 */
 	public function get_tables()
 	{
-		$sql = "SELECT name FROM sqlite_master WHERE type='table' ORDER BY name";
+		$sql = "SHOW TABLES";
 		foreach($this->pdo->query($sql)->fetchAll(PDO::FETCH_NUM) as $table)
 		{
 			if( ! in_array($table[0], DBManager::$hidden))
@@ -50,7 +49,7 @@ class SQLite extends Driver {
 
 		try
 		{
-			$this->table_info = $this->pdo->query("pragma table_info($table)")->fetchAll(PDO::FETCH_ASSOC);
+			$this->table_info = $this->pdo->query("SHOW FULL COLUMNS FROM `$table`")->fetchAll(PDO::FETCH_ASSOC);
 		}
 		catch (Exception $e)
 		{