+++ layla/start.php
@@ -1,11 +1,11 @@
 <?php
-
 Autoloader::namespaces(array(
-	'Components\\Layla' => __DIR__
+	'Layla' => __DIR__
 ));
 
 // --------------------------------------------------------------
 // Set Aliases
 // --------------------------------------------------------------
-Autoloader::alias('Components\\Layla\\Notification', 'Notification');
-Autoloader::alias('Components\\Layla\\HTML', 'HTML');
+Autoloader::alias('Layla\\Notification', 'Notification');
+Autoloader::alias('Layla\\HTML', 'HTML');
+//Autoloader::alias('Layla\\Module\\Form', 'Form');