+++ thirdparty/assetcompressor/asset.php
@@ -14,7 +14,7 @@ class Asset_Container extends Laravel_Asset_Container {
 	{
 		parent::__construct($name);
 		
-		$this->config = Config::get('layla_thirdparty_assetcompressor::assetcompressor');
+		$this->config = Config::get('thirdparty_assetcompressor::assetcompressor');
 		$this->config['cache_dir'] = $this->config['cache_dir'] . '/';
 		$this->config['cache_dir_path'] = path('public').$this->config['cache_dir'];