+++ changes.md
@@ -1,5 +1,6 @@
 #### Bootsparks
-- Added Form alias
+- Bootsparks extends Layla\Form\Renderers\Renderer which is used for removing htmlentities 
+in some of Laravel's Form methods, this allows us to nest them more easily
 
 
 #### Authority